+++ webview-ui/vite.config.ts
@@ -13,5 +13,6 @@ export default defineConfig({
         assetFileNames: `assets/[name].[ext]`,
       },
     },
+    sourcemap: true
   },
 });